requirements.txt
streamlit 
faiss-cpu
sentence_transformers
nltk
emoji